♻️ refactor [#58] : order api refactoring * order 구조 변경
