+++ src/main/java/com/sparta/spring_deep/_delivery/admin/order/OrderAdminRepository.java
@@ -1,6 +1,6 @@
 package com.sparta.spring_deep._delivery.admin.order;
 
-import com.sparta.spring_deep._delivery.domain.order.Order;
+import com.sparta.spring_deep._delivery.domain.order.model.Order;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
