+++ src/main/java/com/sparta/spring_deep/_delivery/admin/order/OrderAdminRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.sparta.spring_deep._delivery.admin.order;
 
-import com.sparta.spring_deep._delivery.domain.order.OrderResponseDto;
+import com.sparta.spring_deep._delivery.domain.order.dto.response.OrderResponseDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
