+++ src/main/java/com/sparta/spring_deep/_delivery/admin/order/OrderAdminRepositoryCustomImpl.java
@@ -4,7 +4,7 @@ import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.DateTimePath;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.sparta.spring_deep._delivery.domain.order.OrderResponseDto;
+import com.sparta.spring_deep._delivery.domain.order.dto.response.OrderResponseDto;
 import com.sparta.spring_deep._delivery.domain.order.QOrder;
 import java.time.LocalDateTime;
 import java.util.List;
