+++ src/main/java/com/sparta/spring_deep/_delivery/admin/order/orderItem/OrderItemAdminRepository.java
@@ -1,7 +1,7 @@
 package com.sparta.spring_deep._delivery.admin.order.orderItem;
 
 
-import com.sparta.spring_deep._delivery.domain.order.orderItem.OrderItem;
+import com.sparta.spring_deep._delivery.domain.order.model.OrderItem;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
