--- src/main/java/com/sparta/spring_deep/_delivery/domain/order/OrderStatusEnum.java
@@ -1,9 +1,0 @@
-package com.sparta.spring_deep._delivery.domain.order;
-
-public enum OrderStatusEnum {
-    PENDING,
-    CONFIRMED,
-    DELIVERED,
-    FAILED,
-    CANCELLED
-}
