+++ src/main/java/com/sparta/spring_deep/_delivery/domain/order/dto/OrderSearchDto.java
@@ -1,5 +1,6 @@
-package com.sparta.spring_deep._delivery.domain.order;
+package com.sparta.spring_deep._delivery.domain.order.dto;
 
+import com.sparta.spring_deep._delivery.domain.order.OrderStatusEnum;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
