+++ src/main/java/com/sparta/spring_deep/_delivery/domain/order/repository/OrderRepositoryCustom.java
@@ -1,5 +1,7 @@
-package com.sparta.spring_deep._delivery.domain.order;
+package com.sparta.spring_deep._delivery.domain.order.repository;
 
+import com.sparta.spring_deep._delivery.domain.order.dto.OrderSearchDto;
+import com.sparta.spring_deep._delivery.domain.order.dto.response.OrderResponseDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
