+++ src/main/java/com/sparta/spring_deep/_delivery/domain/payment/Payment.java
@@ -1,7 +1,7 @@
 package com.sparta.spring_deep._delivery.domain.payment;
 
 import com.sparta.spring_deep._delivery.common.BaseEntity;
-import com.sparta.spring_deep._delivery.domain.order.Order;
+import com.sparta.spring_deep._delivery.domain.order.model.Order;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
