+++ src/main/java/com/sparta/spring_deep/_delivery/domain/payment/PaymentService.java
@@ -2,8 +2,8 @@ package com.sparta.spring_deep._delivery.domain.payment;
 
 import static com.sparta.spring_deep._delivery.util.AuthTools.ownerCheck;
 
-import com.sparta.spring_deep._delivery.domain.order.Order;
-import com.sparta.spring_deep._delivery.domain.order.OrderRepository;
+import com.sparta.spring_deep._delivery.domain.order.model.Order;
+import com.sparta.spring_deep._delivery.domain.order.repository.OrderRepository;
 import com.sparta.spring_deep._delivery.domain.user.entity.User;
 import com.sparta.spring_deep._delivery.domain.user.entity.UserRole;
 import com.sparta.spring_deep._delivery.exception.InternalServerException;
