+++ src/main/java/com/sparta/spring_deep/_delivery/domain/review/model/Review.java
@@ -1,7 +1,7 @@
 package com.sparta.spring_deep._delivery.domain.review.model;
 
 import com.sparta.spring_deep._delivery.common.BaseEntity;
-import com.sparta.spring_deep._delivery.domain.order.Order;
+import com.sparta.spring_deep._delivery.domain.order.model.Order;
 import com.sparta.spring_deep._delivery.domain.user.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
