+++ src/test/java/com/sparta/spring_deep/_delivery/domain/order/OrderIntegrationTest.java
@@ -19,9 +19,11 @@ import com.sparta.spring_deep._delivery.domain.address.entity.Address;
 import com.sparta.spring_deep._delivery.domain.address.repository.AddressRepository;
 import com.sparta.spring_deep._delivery.domain.menu.Menu;
 import com.sparta.spring_deep._delivery.domain.menu.MenuRepository;
-import com.sparta.spring_deep._delivery.domain.order.orderDetails.OrderDetailsRequestDto;
-import com.sparta.spring_deep._delivery.domain.order.orderDetails.OrderDetailsResponseDto;
-import com.sparta.spring_deep._delivery.domain.order.orderItem.OrderItemDto;
+import com.sparta.spring_deep._delivery.domain.order.dto.request.OrderDetailsRequestDto;
+import com.sparta.spring_deep._delivery.domain.order.dto.response.OrderDetailsResponseDto;
+import com.sparta.spring_deep._delivery.domain.order.dto.OrderItemDto;
+import com.sparta.spring_deep._delivery.domain.order.model.Order;
+import com.sparta.spring_deep._delivery.domain.order.repository.OrderRepository;
 import com.sparta.spring_deep._delivery.domain.restaurant.Restaurant;
 import com.sparta.spring_deep._delivery.domain.restaurant.RestaurantRepository;
 import com.sparta.spring_deep._delivery.domain.restaurant.restaurantAddress.RestaurantAddress;
