+++ src/test/java/com/sparta/spring_deep/_delivery/testutil/TestEntityCreateTools.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.spring_deep._delivery.domain.address.entity.Address;
 import com.sparta.spring_deep._delivery.domain.ai.Ai;
 import com.sparta.spring_deep._delivery.domain.menu.Menu;
-import com.sparta.spring_deep._delivery.domain.order.Order;
-import com.sparta.spring_deep._delivery.domain.order.orderItem.OrderItem;
+import com.sparta.spring_deep._delivery.domain.order.model.Order;
+import com.sparta.spring_deep._delivery.domain.order.model.OrderItem;
 import com.sparta.spring_deep._delivery.domain.payment.Payment;
 import com.sparta.spring_deep._delivery.domain.payment.Payment.PaymentStatusEnum;
 import com.sparta.spring_deep._delivery.domain.restaurant.Restaurant;
